More 64-bit type updates

+++ trunk/Source/Clients/Spreadsheets/GDataEntryWorksheet.h
@@ -1,17 +1,17 @@
 /* Copyright (c) 2007 Google Inc.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 //
 //  GDataEntryWorksheet.h
@@ -23,7 +23,7 @@
 #undef _EXTERN
 #undef _INITIALIZE_AS
 #ifdef GDATAENTRYWORKSHEET_DEFINE_GLOBALS
-#define _EXTERN 
+#define _EXTERN
 #define _INITIALIZE_AS(x) =x
 #else
 #define _EXTERN extern
@@ -40,11 +40,11 @@ _EXTERN NSString* const kGDataCategoryWorksheet _INITIALIZE_AS(@"http://schemas.
 
 // Hard upper bound of rows and columns, perhaps including
 // empty rows and columns.
-- (int)rowCount;
-- (void)setRowCount:(int)val;
+- (NSInteger)rowCount;
+- (void)setRowCount:(NSInteger)val;
 
-- (int)columnCount;
-- (void)setColumnCount:(int)val;  
+- (NSInteger)columnCount;
+- (void)setColumnCount:(NSInteger)val;
 
 // convenience accessors
 - (GDataLink *)cellsLink;
@@ -53,4 +53,3 @@ _EXTERN NSString* const kGDataCategoryWorksheet _INITIALIZE_AS(@"http://schemas.
 // listLink is deprecated in V2; use worksheetsFeedURL instead
 - (GDataLink *)listLink;
 @end
-
